change all positions and velocities to float

+++ Engine/Components.h
@@ -8,29 +8,29 @@ class Object {
 public:
   Object();
   Object(SDL_Texture *texture);
-  Object(SDL_Texture *texture, const int x, const int y);
+  Object(SDL_Texture *texture, const float x, const float y);
   Object(SDL_Texture *texture, const SDL_Rect &rect);
 
 public:
   void setTexture(SDL_Texture *texture);
-  void setPos(const int x, const int y);
+  void setPos(const float x, const float y);
   void render(SDL_Renderer *renderer);
   virtual void update(const int dt_ms);
-  void updatePos(const int x, const int y);
-  void setVelocity(const int vx, const int vy);
-  void setVelocityX(const int vx);
-  void setVelocityY(const int vy);
-  int getPosX() const { return m_pos.x; }
-  int getPosY() const { return m_pos.y; }
-  int getVelocityX() const { return m_vx; }
-  int getVelocityY() const { return m_vy; }
+  void updatePos(const float x, const float y);
+  void setVelocity(const float vx, const float vy);
+  void setVelocityX(const float vx);
+  void setVelocityY(const float vy);
+  float getPosX() const { return m_pos.x; }
+  float getPosY() const { return m_pos.y; }
+  float getVelocityX() const { return m_vx; }
+  float getVelocityY() const { return m_vy; }
   void scale(const float factor);
 
 private:
   SDL_Texture *m_texture;
   SDL_Rect m_pos;
-  int m_vx = 0;
-  int m_vy = 0;
+  float m_vx = 0;
+  float m_vy = 0;
 };
 
 class Player : public Object {
@@ -48,8 +48,8 @@ private:
 class Projectile : public Object {
 public:
   Projectile();
-  Projectile(SDL_Texture *texture, const int x, const int y, const int vx,
-             const int vy, const int lifespan_ms, const int damage);
+  Projectile(SDL_Texture *texture, const float x, const float y, const float vx,
+             const float vy, const int lifespan_ms, const int damage);
 
 public:
   void update(const int dt_ms);
